move tooltip annotation to components module
